+++ LibSrc/GlkLibrary.m
@@ -70,7 +70,7 @@ static GlkLibrary *singleton = nil;
 }
 
 - (void) dealloc {
-	NSLog(@"GlkLibrary dealloc %x", self);
+	NSLog(@"GlkLibrary dealloc %x", (unsigned int)self);
 	if (singleton == self)
 		singleton = nil;
 	self.gameid = nil;
